+++ src/Core/WorkSpace/src/Components/QuickFilter/FilterGroup/index.ts
@@ -1,0 +1,1 @@
+export * from './FilterGroup';
