+++ src/Core/WorkSpace/src/index.ts
@@ -4,3 +4,4 @@ export * from './WorkSpaceApi/useWorkSpace';
 export type { TableOptions, TreeOptions } from './WorkSpaceApi/workspaceState';
 export * from './WorkSpaceApi/WorkSpaceTypes';
 
+export * from './Components';
