+++ src/modules/powerBI/src/Components/Filter/PopoverGroup/index.ts
@@ -1,0 +1,2 @@
+export * from './PopoverGroup';
+export * from './VirtualList';
