+++ src/modules/powerBI/src/Components/Filter/index.ts
@@ -1,2 +1,3 @@
 export * from './PowerBIFilter';
 export * from './FilterItems';
+export * from './PopoverGroup';
